Modificacion de la clase servicios conviertendola en un decimal

+++ MyReserve/Models/TablasBBDD/Servicios/Servicios.cs
@@ -2,7 +2,7 @@
     public class Servicios {
         public int ser_id {  get; set; }
         public string? ser_nombre { get; set; }
-        public int ser_precio { get; set; }
+        public decimal ser_precio { get; set; }
         public string? ser_pelu_id_fk { get; set; }
         public string? ser_cat_id_fk { get; set; }
         public int ser_asociado { get; set; }
